Remove BestRobotDrive.java since the new system no longer uses it.
It was in a state of half conversion to an 8 motor system and was a horrible
mess in the first place so I decided it was no longer necessary. If I need to
reference the algorithms, I can just look at another branch. I also removed one
last accidental reference to BestRobotDrive that remained in the new system.